Remove superfluous comment from justfile
GitHub Linguist is reporting this repository as a justfile repo,
despite its primary source being Rust.  It's because the justfile is
just a few lines longer than the main Rust file.  I can't see any way
to include justfile in the statistics while excluding it as the main
language.  Instead of figuring that out (if it is even possible), it
is easier to remove some of the copypasta of the justfile that I have
taken from my configs, which has been alternatively linked
instead (for the curious reader).